Move remarkable renderer to it's own file
This cleans up the text reducer some
